requirements.txt
langchain
langchain-core
langchain-community
flask
python-dotenv
langchain_google_genai